Move all the test listeners to the testing module
